+++ src/App/App.js
@@ -33,7 +33,7 @@ const App = (props) => (
 		</Switch>
 	</Router>
 );
-import AppStateDecorator from './AppStateDecorator'
+//import AppStateDecorator from './AppStateDecorator'
 export default MoonstoneDecorator(
-	//App);
-	AppStateDecorator(App));
+	App);
+	//AppStateDecorator(App));
