+++ src/App/App.module.less
@@ -1,3 +1,4 @@
-.app {
+.panel {
 	// styles can be put here
+	display: block;
 }
