+++ src/components/Common/BoxContainer.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const BoxContainer = (props) => {
+    return (
+        <div className="container">
+            
+        </div>
+    )
+}
+export default BoxContainer;
