+++ src/components/Common/BoxHeader.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import { Header } from '@enact/moonstone/Panels';
+import BtnGoBack from './BtnGoBack';
+import BtnGoMain from './BtnGoMain';
+
+const BoxHeader = ({ title, ...props }) => {
+    return (
+        <Header title={title}>
+            <BtnGoBack/>
+            <BtnGoMain/>
+        </Header>
+    )
+}
+
+export default BoxHeader;
