+++ src/components/Common/index.js
@@ -1,0 +1,8 @@
+export {default as BoxContainer} from './BoxContainer';
+export {default as BoxHeader} from './BoxHeader';
+export {default as BoxPanel} from './BoxPanel';
+export {default as BtnGoBack} from './BtnGoBack';
+export {default as BtnGoMain} from './BtnGoMain';
+export {default as BtnLogo} from './BtnLogo';
+export {default as StayTime} from './StayTIme';
+export {default as Video } from './Video'
