+++ src/components/Kennel/VomitCheck.js
