+++ src/components/Location/index.js
@@ -1,0 +1,1 @@
+export {default as KakaoMap} from './KakaoMap'
