+++ src/components/Main/index.js
@@ -1,0 +1,4 @@
+export {default as BoxMainBtn } from './BoxMainBtn';
+export {default as BtnCheck } from './BtnCheck';
+export {default as BtnConnect } from './BtnConnect';
+export {default as BtnPush } from './BtnPush';
