+++ src/components/README.md
@@ -1,3 +1,13 @@
 Reusable components for your application go here<br>
 View에 들어갈 components들<Br>
 action이랑 connect 한 container 포함
+
+Common/에는 공통적으로 사용되는 component
+
+Home,Kennel,Location,Main/에는 각각 View에 들어갈 component들이고
+
+Common에서 받아서 변형하고 함수 등을 추가하여 사용
+
+각각 폴더 안에는 index.js를 넣어 import 하기 편하게함
+
+폴더 생성은 최소한으로
