+++ src/views/HomeView.js
@@ -1,14 +1,13 @@
 import Button from '@enact/moonstone/Button';
 import { Panel, Header } from '@enact/moonstone/Panels';
 import React from 'react';
-import BtnGoMain from '../components/Buttons/BtnGoMain'
+import { BoxHeader,BoxPanel } from '../components/Common';
 
 const HomeView = (props) => {
     return (
-        <Panel >
-            <Header title="스마트 홈" />
-            <BtnGoMain/>
-        </Panel>
+        <BoxPanel >
+            <BoxHeader title="스마트 홈" />
+        </BoxPanel>
     )
 }
