+++ src/views/Kennel/LampView.js
@@ -2,17 +2,15 @@ import Button from '@enact/moonstone/Button';
 import { Panel, Header } from '@enact/moonstone/Panels';
 import Slider from '@enact/moonstone/Slider';
 import React from 'react';
-import BtnGoBack from '../../components/Buttons/BtnGoBack';
-import Video from '../../components/Video';
+import { BoxPanel, BoxHeader,Video } from '../../components/Common';
 
 const LampView = (props) => {
     return (
-        <Panel>
-            <Header title="무드등 조절" />
-            <BtnGoBack/>
+        <BoxPanel>
+            <BoxHeader title="무드등 조절" />
             <Slider/>
             <Video/>
-        </Panel>
+        </BoxPanel>
     );
 }
