+++ src/views/README.md
@@ -1,1 +1,9 @@
 Composite components that make up a distinct view go here
+
+View는 component에서 받은것을 조합하는 역할
+
+따라서 style, 함수 등의 적용을 최소화하고
+
+style을 바꾸고 싶다면 component를 만들던가, className을 이용하고
+
+함수를 적용하고 싶으면 component 사용
